requirements.txt
autopep8==1.5.4
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.21.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.1
python-dotenv==0.15.0
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
